Refactor seed test code into its own helper function and update test
